// frontend/src/backend.d.ts
export interface CheckoutRequest {
    paymentMethod: PaymentMethod;
    cancelUrl: string;
    buyerId: string;
    items: Array<ShoppingItem>;
    successUrl: string;
}
// ...
export interface LicensingOption {
    id: string;
    terms: string;
    name: string;
    description: string;
    price: bigint;
}
// ...
export interface ShoppingItem {
    productName: string;
    currency: string;
    quantity: bigint;
    priceInCents: bigint;
    productDescription: string;
}
// ...
export enum MediaType {
    music = "music",
    video = "video",
    text = "text",
    image = "image"
}
export enum PaymentMethod {
    stripe = "stripe",
    merchantServices = "merchantServices",
    nationalBankcard = "nationalBankcard"
}